Upgrade NodeJS version and switch to custom storybook
