+++ .nvmrc
@@ -1,1 +1,1 @@
-16.14.2
+20.15.1
