--- .storybook/main.js
@@ -1,3 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.ts'],
-};
