+++ package.json
@@ -26,11 +26,10 @@
     "lint": "eslint ./src/**/*.ts",
     "prepush": "npm run lint && npm run test && prettier --check ./src/*",
     "prepublishOnly": "npm run build && npm run prepush",
-    "storybook": "start-storybook",
-    "build-storybook": "build-storybook -c .storybook -o out"
+    "storybook": "vite dev",
+    "build-storybook": "vite build --outDir out"
   },
   "devDependencies": {
-    "@storybook/html": "^6.1.11",
     "@types/jest": "^26.0.19",
     "@typescript-eslint/eslint-plugin": "^4.11.1",
     "@typescript-eslint/parser": "^4.11.1",
@@ -41,7 +40,8 @@
     "jest": "^26.6.3",
     "prettier": "1.19.1",
     "ts-jest": "^26.4.4",
-    "typescript": "^4.1.3"
+    "typescript": "^4.1.3",
+    "vite": "^5.3.4"
   },
   "husky": {
     "hooks": {
