+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Camera } from './types';
+export type { Camera } from './types';
 export { default as drawMethods } from './drawMethods';
 export { default as objectTypes } from './objectTypes';
 export { default as createDrawFunction } from './createDrawFunction';
