+++ src/storybook/stories/import-png.d.ts
