+++ src/storybook/stories/resources/js-logo.png
